+++ release_notes.md
@@ -10,3 +10,4 @@
 - Update OpenTelemetry instrumentation and exporter packages (#11766)
 - Skip `WebJobsStorageHealthCheck` when no active script host is available (#11791)
 - Avoid false-positive connection string checks when constructing `BlobServiceClient` (#11794)
+- Fix .NET10 breaking change with connection-string prefixed env vars. (#11793)